Infrastructure: report Qt version & Windows WOW details in About dialog (#7273)
#### Brief overview of PR changes/additions
Detect and report whether a Windows builds of Mudlet is running natively
on a 64-Bit or 32-Bit CPU or is a 32-Bit version running in the WOW64
sub-system in 64-Bit Windows. Also give the version details of the Qt
libraries and if the run-time version is not the same as that for
compilation give both.

#### Motivation for adding to Mudlet
Easier identification of the system involved in bug reporting,
especially with the probable move to Qt6 in the future.

#### Other info (issues closed, discussion etc)

---------

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>
(cherry picked from commit 4054217f16a9813b8cc2f322d96dca2bbc854d79)