Improve: move 'report codepoints errors' under a compile-time option (#7332)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
Move 'report codepoints errors' under a compile-time option,
`DEBUG_UTF8_PROCESSING`, for developers to use as they see fit when they
are debugging codepoint errors.

While the original idea was that ordinary Mudlet users would report
these codepoint errors to Mudlet Makers, the feature since its
introduction 5 years ago has not yielded that many fixes. It did cause
some users to be confused upon seeing the message and unfortunately it
is triggering a crash that we are not able to pinpoint and fix.
#### Motivation for adding to Mudlet
Fix https://github.com/Mudlet/Mudlet/issues/6811 which has been around
for a while and is heavily affecting a user in
https://discord.com/channels/283581582550237184/1260071395279503381
today.
#### Other info (issues closed, discussion etc)
In retrospect enrolling ordinary Mudlet users into what really is a
developer-oriented, highly technical reporting program didn't pay
dividends and we should keep that in mind in the future for any other
such initiatives.