improve: Consolidate profile loading logic (#7395)
/claim #7334
/claim #7357

#### Brief overview of PR changes/additions

- Consolidates profile loading logic for both GUI dialogue loads and CLI
loads in one function
- fixes bugs caused by predefined games not loading settings correctly
in certain cases

#### Motivation for adding to Mudlet

Gets rid of errors about an empty URLs/ports and possibly unwanted SSL
connections.

#### Other info (issues closed, discussion etc)

I decided to go with the approach of unconditionally loading settings
for predefined games from `TGameDetails` because this would fix all
existing cases of "broken" profiles, as opposed to the approach I
suggested earlier, which would only have prevented further broken
profiles being created.

Unifying the duped logic was a bit intimidating seeing as most of those
lines haven't been touched in a decade or so, but I tried to keep actual
changes in what is being executed minimal. Though I did find a few
things that were being done in only one branch but not the other, which
seemed to me like they should have really been in both, and them not
being so was a bug.

I'm actually not 100% sure if this solves #7357 (I can't reproduce it;
config dumps from someone who can would be great), but logically this
change _should_ fix any issues with loading settings for default games.