Fix: Mudlet crashing when changing mapper room symbol font with no mapper
#### Brief overview of PR changes/additions
Checks that the pointer to the font that is used for the room symbols in
the 2D Mapper is non-null before it is dereferenced in the profile settings
dialogue.

#### Motivation for adding to Mudlet
If there is not a mapper present - because the end-user or the packages
they are using have not yet created one then trying to change the font that
*will* be used for this purpose later on will immediately cause the Mudlet
application to crash.

#### Other info (issues closed, discussion etc)
Should close #7346.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>