Fix: allow revised addRoom(...) to work without an area as second argument
#### Brief overview of PR changes/additions
Moves a line of code that was clobbering the Lua <==> C argument passing
"stack" (it was outputting a boolean value and overwriting a value that
was needed as an input for code later in the function) so that it now acts
after the inputting parts are done.

#### Motivation for adding to Mudlet
To fix a defect introduced by #7184 which meant that not providing a second
argument to the revised `addRoom(...)` function meant it read a `true`
value that it had itself placed into the argument stack.

#### Other info (issues closed, discussion etc)
Also fixed an more minor issue where providing an invalid area still seemed
to work correctly in that a `true` value was returned to the caller. The
room was and now still is created but it now causes a nil+error message
to be returned. The room would instead be placed in the "Default Area" -
HOWEVER deferring the recalculations for that area seems to hide the room
in the map - it becomes a "ghost" that does have the right coordinates and
area number (as revealed by `getRoomCoordinates(...)` and `getRoomArea(...)`
but it does not show in the mapper until the map is saved and reloaded (or
the Lua `auditAreas()` function is used - which happens during a map load).

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>