Fix: correct QSignalMapper usage for Qt6
#### Brief overview of PR changes/additions
Revise some signal names to handled changes made in Qt 5.15 and demanded
(because old named signals have been removed) in Qt 6, for `QSignalMapper`.

#### Motivation for adding to Mudlet
Prevent error messages and problems with a signal-slot connection not
working for user-actions added to the 2D mapper context menu and the
fore-/back-ground colour selection buttons in colour triggers.

#### Other info (issues closed, discussion etc)
`QSignalMapper` has been depreciated in Qt6 as the functionality it proved
can now be done in C++ lamda functions - however making such changes
requires a firm understanding of how each of them works and in the meantime
it is less complex to just use the functor form of the `connect(...)` - to
replace the obsolete `const char[]` form and to use the newer names that
are supported from Qt version **5.15.0**.

I pointed out this issue in our Discord "#mudlet-development" channel at
[2024/08/03 21:16 UTC](https://canary.discord.com/channels/283581582550237184/283582439002210305/1280637687552020510).

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>