Fix: remove memory leak from use of QSignalMapper (#7424)
#### Brief overview of PR changes/additions
The context menu for the 2D mapper uses a `QSignalMapper` that is
allocated off the heap with `new` however it is parented to the `T2DMap`
class instance and is never explicitly `delete`d. As such, instances
build up during the lifetime of a profile (one for every time a
right-click on the 2D mapper generates a context menu). This PR fixes
that by parenting the mapper to the menu itself which is itself
short-term and does not persist longer than necessary.

#### Motivation for adding to Mudlet
Avoiding a memory leak that accumulates during the life-time of the
mapper widget in a profile.

#### Other info (issues closed, discussion etc)
Also in passing I spotted a near-by `QString` being compared against a
raw C empty string - which I fixed because we have a coding style that
says we should be using `QString::isEmpty()` to better state the code
intent and to avoid raw string literals.

---------

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>