Add: ability to arrange and store profiles' order in connection dialog
#### Brief overview of PR changes/additions
Modifies the connect profiles dialog so that the icons are no longer forced
to be in alphabetical order. Adds a context menu option to the widget that
displays the profile icons so that the automatic layout in a grid can be
switched off so that icons can be dragged/rearranged and then re-enabled
so the icons return to a grid layout in the new order which is then
preserved between invocations of the dialogue. The order is saved in the
main Mudlet `QSetting`s file so it also persists between Mudlet application
runs. The size that the user has made the connection dialog is also saved
so that it can be resized back to that the next time it is used.

#### Motivation for adding to Mudlet
To ease the management of profiles' in the connection dialog.

#### Other info (issues closed, discussion etc)
* Tweaks the minimum dimensions of the widget that shows the profile icons
so that on startup it can support 3 icons across.
* Reduces to a minimum (1) the point-size of the font for the (unused) text
that would appear under the icons so as to reduce the vertical separation
between them to the smallest they can be.
* Polishes the `connection_profiles.ui` file so it looks/behaves a bit
better - including adjusting some size settings and constraints so that
it can show three columns of profile icons when it is first used. Given
that the size is now also saved it means that the user can arrange the
icons in the order they wish and if they have the space they can show more
of them at once.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>