Fix: only remember predefined MUDs that get deleted
The action of "Removing" a profile from the "Connect profiles" dialogue
was remembering the name of ALL such profiles that were discarded. However
this meant that if the name was reused and recreated and it was NOT a
predefined profile it would never show up in the dialogue again. This
commit now only "remembers" if a predefined MUD was removed.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>