Improve: insert ALL profiles into QListWidget but hide unwanted ones
This is so that the stored "order" always works even when Mudlet is invoked
with command line options to limit the predefined profiles or to load a
particular. In particular this includes the "Mudlet self-test" one. During
development/testing of this PR I was unexpectedly seeing this profile even
though it was supposed to be hidden (it turns out that
`QListWidgetItem::setHidden(bool)` does not work on an item that is NOT
inserted into a `QListWidget` - it forgets that it is hidden; I will be
filing a QTBUG about this when I can)! So as to make this more obvious and
to close #6443 I've made and included an icon for this profile even though
it should not be seen currently.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>