Improve: handling of full-screen, maximise and minimise application
#### Brief overview of PR changes/additions
Improves the handling of the state of the main application window by better
handling of the Qt::WindowsStates flags. It also makes the button that is
used to put Mudlet into full-screen mode respond to external OS/DE controls
that could put the application into that mode as well. That button is now
also a "checkable" one with an icon that changes to reflect the state
Mudlet is in.

#### Motivation for adding to Mudlet
I've been unhappy for a while about how Mudlet handles transitions to/from
full-screen mode in that it was using `showNormal()`/`showMaximised()`/
`showFullScreen()` methods without considering the other `Qt::WindowStates` flag.

#### Other info (issues closed, discussion etc)
Also adds correct icons to Full-screen button which is made to be clickable
in order to be able to show either the standard "view-fullscreen" or
"view-restore" icons (which I sourced from the KDE Oxygen set).

This should close #1925.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>