Fix: adapt this PR to accomodate parallel changes made by #7152
Extended the functions disabled by the login timeout to include the
`sendCharacterName()` function - and allow the timer that provides that
timeout to be stopped by the GMCP authenticator that #7152 introduced. Such
a stop also immediately disables those functions so that it will neutralise
any scripts or triggers that would be using them.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>