Fix: stop showing "Playing Mudlet" when Discord is not enabled for a profile
#### Brief overview of PR changes/additions
Disables sending updates to the local Discord application if a profile is
not active or the one that is active does not have the discord integration
enabled.

#### Motivation for adding to Mudlet
This was leaking information about what the user is doing when they may not
have given Mudlet permission to do so. Specifically it stops Mudlet
reporting to Discord that the user is using Mudlet when they have unchecked
the box that should thus have disable Mudlet from talking to Discord via
the RPC mechanism.

#### Other info (issues closed, discussion etc)
This should close #6967.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>