Fix: remove spaces from reply to MXP version tag. (#7144)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
Removes spaces from the version string returned by MXP to avoid parsing
issues in the mud. Also, like other MXP clients, mudlet will now report
the version nummer without prefixing the client name (which is reported
in a separate field).
#### Motivation for adding to Mudlet
Improve MXP compatibility
#### Other info (issues closed, discussion etc)
This is meant as a fix to issue #7140 which popped up in recent PTB
versions when the format of the internal version string was changed.