Infrastructure: Update widechar_width.h to latest upstream (#7449)
#### Brief overview of PR changes/additions
:crown: An automated PR to update the built-in `src/widechar_width.h` to
its latest version in upstream. We don't include it as a submodule as
it's just one file.
#### Motivation for adding to Mudlet
Better emoji and symbol support! This file helps us tell if a particular
character is 0, 1, or 2 widths wide so we can show it correctly in the
display.
#### Other info (issues closed, discussion etc)
_update triggered by refs/heads/development
c084678bed001a8bae93714f78c53b1e4ed9205a_

Co-authored-by: mudlet-machine-account <mudlet-machine-account@users.noreply.github.com>