Add: Quick clear input-line functionality (#7450)
#### Brief overview of PR changes/additions

This adds functionality to handle input line clearing as described in
Issue https://github.com/Mudlet/Mudlet/issues/7177

A user may press DOWN (with no modifiers) to quickly clear what they
have typed in the command line.

The clearing functionality will only trigger if the user is in the
middle of typing a command (nothing *selected* on the command line), and
while not navigating through their command history. This allows history
navigation and autocompletion to function as normal while allowing an
additional feature of quickly clearing the command line.

#### Motivation for adding to Mudlet

Occasionally a player may find themself in the middle of typing a
command when they wish to quickly switch to another command. Previously
the options to clear the command line and start typing again involved
either sending the unwanted command which could cause unwanted game lag
set for the player (for example a backstab command on someone not there
may result in a combat round of lag on a *miss*), or using multiple
keystrokes to select all and clear with backspace.

#### Other info (issues closed, discussion etc)