Fix: include time-stamps in NAWS calculations
#### Brief overview of PR changes/additions
This PR subtracts the number of characters in the time-stamp from the
"width" value that is sent out to the server when the time stamps are
shown.

#### Motivation for adding to Mudlet
For Game Servers that do *Server-side* wrapping of text the details Mudlet
sends out for the NAWS (*Negotiate About Window Size*) fails to account for
the reduction in on-screen width (that is available for the Game text) when
time stamps are showing in the main console.

#### Other info (issues closed, discussion etc)
I spotted this whilst playing After the Plague
(telnet:\\atp.pedia.szote.u-szeged.hu::3000) which does do Server-side
wrapping of the text it sends (which does mess with the generic mapper when
the screen size is small enough to cause the "exits" line to wrap onto a
second line! :rolling_eyes:)

Obviously as Mudlet does not rewrap the main console text when the window
is resized this does mean that long lines that fitted on screen when the
time-stamps are not shown do get some characters pushed off the right side
when they are.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>