Fix: when closing multiple profiles wait for quiescence before proceeding
#### Brief overview of PR changes/additions
Wait for the Mudlet application to not be busy before trying to close
(multiple) profiles - e.g. by holding down the keys set to do the
"Close profile" short-cut.

#### Motivation for adding to Mudlet
Stopping it from going :boom: for end-users.

#### Other info (issues closed, discussion etc)
It was really hard to determine why the crashes were occurring and running
Mudlet inside the Qt Creator debugger (`gdb`) didn't positively give an
answer. The problems seem to be because a `Host` pointer wasn't working
properly - according to the debug what was pointed to was still valid (the
values in the pointed to `Host` instance looked reasonable) but the actual
de-referencing was failing inside Qt library code. At that point there was
a stack of several hundred nested/repeating frames for each of the multiple
profiles that were in the process of being closed so that is what inspired
me to reorganise the process to not call `mudlet::closeHost(const QString&)`
directly from `mudlet::slot_closeProfileRequested(int)` but instead do it
via a "zero-timer" which will wait for pending stuff to be done before
trying to close the next profile. This is not IMVHO a perfect solution
because the actual cause of the problem has not been precisely determined,
however it does seem to be "good enough" providing other testers confirm
the improvement I think this makes.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>