Fix: revert broken indentation (#7441)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
This reverts commit 8d83f7877b5c888b2cb02456842dc5ddd7fa2223 / PR
https://github.com/Mudlet/Mudlet/pull/7039.
#### Motivation for adding to Mudlet
Fixes https://github.com/Mudlet/Mudlet/issues/7316
#### Other info (issues closed, discussion etc)
This is a high-level issue that has been affecting many players for many
months now, and unfortunately it has no traction in getting fixed. It is
also holding up a release. Lets revert it so we can do a release with
all of the other fixes that have been done.

The bounty for the original issue was already paid out and unfortunately
the author is unresponsive in providing a fix, so we'll just deduct the
incorrectly paid out bounty amount in a future bounty.

Co-authored-by: Vadim Peretokin <vadi2@users.noreply.github.com>