Add: extend GUI Languages to support Hebrew (#7457)
#### Brief overview of PR changes/additions
Adds a new language to the selection available on the first tab of the
preferences.

#### Motivation for adding to Mudlet
Extends potential locale support for Mudlet to include Israel.

#### Other info (issues closed, discussion etc)
This PR comes about following a discussion between myself, @vadi2 and a
speaker of Hebrew in the Mudlet's **#translation** Discord channel on
2024/10/04 18:50 (UTC):

https://www.discord.com/channels/283581582550237184/603289809116725248/1291834973485142158
**NOTE: Hebrew script is a Right-to-Left one and selecting it will cause
the GUI to reverse. Some quick testing has suggested that this might
generate some oddities in our UI design that will need addressing,
particularly where the GUI had assumed (or encoded - via "alignment"
settings a Left-to-Right basis - the Qt framework is intelligent to swap
things around correctly in the opposite case but only if the correct
"flags" are used - i.e. ones that automagically swap over when the
layout direction is reversed.) See:
https://doc.qt.io/qt-6/qt.html#AlignmentFlag-enum particularly to make
use of: `Qt::AlignLeading` and `Qt::AlignTrailing` instead of
`Qt::AlignLeft` and `Qt::AlignRight`.**

This PR includes a couple of "dummy files":
* translations/translated/mudlet_he_IL.ts
* translations/lua/translated/mudlet-lua_he_IL.json

these are expected to be overwritten by the first upload of proper
translations from CrowdIn to include the `he_IL` {Hebrew (Israel)}
locale which have already been instigated on that platform by @vadi2.
The first of these is a copy of the source `translations/mudlet.ts` with
no changes except the target language marker, and the second is a quick
and probably faulty set of translations for the Lua sub-systen which I
cobbled together with Google translate and which is probably wrong
anyhow.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>