Fix: Use Apple Emoji on MacOS (#7452)
#### Brief overview of PR changes/additions
Mudlet built with Qt5 on MacOS doesn't correctly fall back to using
Apple Color Emoji for either user-selected or default system fonts.
This change adds Apple Color Emoji to the font family of both the
user-selected font and the default system fonts, allowing for correct
emoji display. This is especially visible with compound emoji that use
ZWJ.

As far as I can tell, this issue is greatly improved or non-existent in
Qt6 on MacOS, so I've put these changes behind a QT_VERSION_CHECK for
that reason.

#### Motivation for adding to Mudlet
I would like emoji to render nicely on the Mac. 


#### Other info (issues closed, discussion etc)

Co-authored-by: Vadim Peretokin <vperetokin@hey.com>