Fix: update console after clearWindow(...)
#### Brief overview of PR changes/additions
Clean up the distribution of code associated with clearing a `TConsole`
amongst the classes and add extra calls to methods that remove the display
of the lower `TTextEdit` (i.e. remove the split) and ensure that the upper
one is redrawn after the contents have been removed.

#### Motivation for adding to Mudlet
This - not clearing the lower pane and the split even though all the data
has gone so there is no need for the split - has been an issue (though not
recorded as one) for a long time as far as I can determine.

#### Other info (issues closed, discussion etc)

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>