Fix: repair missing libraries for 32 bit Windows builds
#### Brief overview of PR changes/additions
Add three libraries to those included in the packaged application for
32-Bits Windows builds. They not longer seem to be detected by the `ntldd`
tool (even though the 64-Bits one does does so) - whether the loading
mechanism has changed so that they are loaded dynamically now so no longer
show up in the 32-Bit build seems a possible but weird explaination given
that some testing I've done with 64 Bit builds in both Qt 5 and 6 do show
them.

#### Motivation for adding to Mudlet
This is essential for getting 32-Bit builds working again after
https://github.com/msys2/MINGW-packages/pull/22130 restored two Qt5
packages that had gotten dropped.

#### Other info (issues closed, discussion etc)
This was first reported on Discord by **@bcross32** in the Discord
**[#mudlet-testing](https://www.discord.com/channels/283581582550237184/427919962561052673/1293540881818845215)**
channel at 2024/10/09 11:49 UTC and also by **@karahd** (gesslar) in the
**[#help](https://canary.discord.com/channels/283581582550237184/283582068334526464/1293753289116221543)**
one at 2024/10/10 01:53 UTC.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>