Improve: make main menu items be disabled the same as main toolbar buttons
#### Brief overview of PR changes/additions
Adds code to enable or disable individual main menu items at the same time
as the corresponding (where there is one) main toolbar buttons.

#### Motivation for adding to Mudlet
For a consistent user interface so that different presentations of the same
control are the same across the UI.

#### Other info (issues closed, discussion etc)
The need for this was observed by **@dt192** in our Discord guild's
**#mudlet-development** channel at **[2024/10/10 14:51 UTC](https://www.discord.com/channels/283581582550237184/283582439002210305/1293949027435544707)**

Note: There has been two IRC menu options for a very large part of Mudlet's
history, one for a per-profile one that can be set to connect to any
channel on the IRC Server (LiberaChat) using a built in (libcommuni)
client; the other calls up a web based client run on that Server's own
web-site and accessed via the end-user's web-browser and it was able to be
used even without a profile selected and it calls our own "channel". As
such the latter was capable of being used at any time, even without a
profile loaded. At some time in the past the two menu items were rearranged
to only connect to the per-profile system. I have reverted that and
reconnected the "Live help chat" item to the web-server one. As a result
of the changes in this PR it meant that when a profile is not loaded onlu
one item in the "Toolbox" menu (for the Live Health Chat) was enabled and
simultaneously the only item in the "Help" menu that was disabled was the
"IRC" option. This looked wrong so I took the opportunity to swap them
around.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>