Improve: remove almost unused multi-line bool from one TTrigger constructor
In the cases where the new argument to specify a multi-line trigger is NOT
provided (old scripts) ensure the behaviour of past Mudlet versions is
exactly reproduced.

Also cleans up some usage of a raw string literal `"a"`, particularly in
places where it isn't needed, as a name for a Mudlet "item".

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>