Revise: make new argument to permXxxxTriggers an integer to convey delta value
In the case where a multi-line trigger IS wanted there is also a need to
provide a number of lines in which all the conditions are met. In the editor
this is done via a 0 or greater spin-box which is enabled when the trigger
is made a multi-line one. Importantly the number is unused for the other
type of trigger (a "multi-item") one. Therefore it seems sensible to change
the design of the orginal PR so that the boolean proposed is changed to
an integer - and so that in the case where a multi-item trigger has the
number of lines extra to match in (the `(int) TTrigger::mConditionLineDelta`
value) being zero (the default) we should use a negative value to signify
a multi-item (a.k.a. an OR-type) trigger. FTR the multi-line trigger is
also been described as an AND-type trigger.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>