Infrastructure: remove unneeded control from trigger editor
#### Brief overview of PR changes/additions
Simplify and improve the "trigger" editor part of the "Editor" by
combining/removing some group boxes and making a third one not have a
check-box. Also move them all from the bottom of the left to the right
side of that area and improve the behaviour as the splitter for the right
side is adjusted.

#### Motivation for adding to Mudlet
Improving the UI - helping to make the controls a bit more intuitive.

#### Other info (issues closed, discussion etc)
Having a QSpinBox to control the "condition line" delta for AND/Multi-line
triggers as well as placing it in a checkable group-box to enable it is
more complex than it needs to be. This PR removes the checkbox and instead
uses the ability of a `QComboBox` to have a "special value" which it
presents different - in this case to report the trigger as an OR/Multi-item
type trigger. Having done this I realised that it and the other two items
at the bottom of the left side of the "triggers" editor could be simplified
further and all moved to the right side. Then I had to tweak the code that
showed/hid the right and bottom widgets (since the bottom ones had gone).
Finally I put in code to disable the "OR multi-item"/"AND multi-line"
control when there is less than two "items" in the trigger.

Also:
* rename `(QWidget*) dlgTriggerEditor::HpatternList` ==>
                                  `dlgTriggerEditor::mpWidget_triggerItems`
* convert the - not used as a button - `pushButton_prompt` from a
`QPushButton` to a `QLabel` - and make it be enabled when GA/EoRs are
detected - as well as changing the text as before.

This was prompted by working on the fix for #7480.