Add: permExactMatchStringTrigger(...) and permColorTrigger(...) functions #### Brief overview of PR changes/additions Adds code/LUA API functions to allow the creation of (permanent) "exact match" and "color trigger" types via Lua scripts/packages.
#### Motivation for adding to Mudlet
Consistency in API with other existing permXxxxTrigger(...) functions.

#### Other info (issues closed, discussion etc)
Theoretically there ought to also be functions for "lua function", "line
spacer" type item based triggers - however the latter of those is
(probably) not helpful on their own which suggests that additional
functions/functionality to add/modify/remove individual "items" to existing
triggers is desirable. Conversely, a separate PR to add a
`permLuaFunctionTrigger(...)` function is now on my TODO list!

The functions added by this PR have been modified before publication to
accept the recently added (by myself) "multi-line delta" argument of PR
#7493 - though these functions post-date that PR this fifth argument is
also considered to be optional so the new functions behave as the existing
ones do.

The new `permColorTrigger(...)` function takes a table of sub-tables as
the third argument. Whilst the "color trigger" type currently only handles
either/both a foreground and a background colour (stored as a string
representation with those elements as a three digit number inside that
string or a couple of "special" string values otherwise) I have medium term
plans to extend the type to handle other text formatting details (even
those that we currently detect but do not reproduce: bold, italic,
underline, blink, etc.) I expect to store this as a series of words, one
for each aspect, appended to that string and possibly preceded with the
prefix "'NOT_'" or the symbol "'!'". So this can be added later to the
Lua API function it will also accept - but issue a `[ WARN ] -` additional
string arguments in the "sub-table" for each trigger "item" after zero, one
(foreground only) or two (foreground and background). This is so that if
a script/package targetting a later Mudlet version that does handle this
is used with a version of Mudlet with the `permColorTrigger(...)` as of
this PR then the user will get a clear warning that it likely won't work as
expected. (Obviously the current and previous Mudlet versions will just
barf on a usage of the non-existing - in them - `permColorTrigger(...)`!)

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>