Infrastructure: make TRoom coordinates private #### Summary of PR Changes/Additions Makes the coordinate members of the `TRoom` class private so that access to them can be tracked via methods to set and get them.
#### Motivation for Adding to Mudlet
This is so that the setters can then subsequently include any extra code
that needs to be aware when the room is moved. I intend to improve the
detection of rooms being placed in the same position but realised this
would be a good preliminary step.

#### Additional Information (related issues, discussions, etc.)
Removes some dead code setting but not using `(int) quads` and
`(int) verts` in `(void) GLWidget::paintGL()`

Also using the mouse to drag and thus move selected rooms when those rooms
were on different levels would squash them all down to be on the same
z-coordinate as the "highlighted centre of the selection" room. This is not
as helpful it might seem and instead increased the likelyhold of causing
room collisions - so now each room will retain it's z coordinate if it is
not on the same level as the centre of the multiple room selection.

Also move code that likely needs to be run whenever rooms are added/removed/
moved within an area to a common block of code (`(void) TArea::clean()) to
help keep things DRY. I intend to put code to update a per area record of
rooms that are in the same place within that block in the future - so that
the record can be reused without having to be repeatedly recalculate,
especially in the paint event for the 2D mapper.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>