Improve: mark overlapping rooms in 2D mapper
#### Summary of PR Changes/Additions
Puts a (user delectable but defaults to yellow) around rooms in the 2D
mapper that are at the same location (collides with) another room to help
map-makers to notice where this happens.

#### Motivation for Adding to Mudlet
It is easy when map making to accidentally put one room on top of another
and otherwise this only shows up when one tries to select one of such rooms
and the room list widget appears in the mapper. This code also works when
dragging rooms around so can be used to avoid dropping the currently
selected room(s) into occupied spaces.

#### Additional Information (related issues, discussions, etc.)
This code overrides the normal room boarders code in that it always
operates even if the user has turned off room boarders normally. It also
overrides the coloured boarder used for selected rooms - but such rooms
also have a gradient background fill provided they are drawn big enough so
it should still be possible to differentiate between the two cases...

The colour chosen is remembered on a per-profile basis.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>