Fix: prevent crash with empty theme details in game save #### Brief overview of PR changes/additions Checks that `dlgTriggerEditor::slot_editorThemeChanged()` is invoked with a valid theme set up and to fall-back to the built in "Mudlet" one if it isn't.
#### Motivation for adding to Mudlet
Prevent crashes caused by faulty saved games where the editor theme
settings held as "attributes" in the `<Host >` element of the XML file
(`mEditorTheme` and `mEditorThemeFile`) were empty strings. The mechanism
that caused this on my PC is not clear but it is worth preventing it
causing problems for other users should they get in the same state.

#### Other info (issues closed, discussion etc)
Ideally I would have liked to have put a `nullptr` check into
`(void) SingleLineTextEdit::setTheme(const QString& themeName)` but I chose
not to do so for two reasons:
1. Any error message would be spammed 50 times, once for each of the fifty
trigger items in each trigger - whether used or not.
2. That file was put into Mudlet under unclear licensing terms (it has the
GPL2+ header text but is missing the Copyright line of the original author,
if I were to apply my Copyright for a contribution in that file it would
look as though I authored the whole file - which is NOT the case!) Until
this is cleared up I am reluctant to touch the file.

Also correct the text in a couple of comments that refer to the file name
of the editor "Mudlet" theme we bundle within the application.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>